Moves content-renderer card into core.
