+++ frontend_build/src/read_bundle_plugins.js
@@ -24,6 +24,7 @@ var libs = function(kolibri_name) {
     'vue': kolibri_name + '.lib.vue',
     'kolibri': kolibri_name,
     'core-base': kolibri_name + '.lib.coreBase',
+    'content-renderer': kolibri_name + '.lib.contentRenderer',
   };
 };
