+++ frontend_build/src/webpack.config.base.js
@@ -80,6 +80,7 @@ var config = {
       'kolibri_module': path.join(__dirname, '..', '..', 'kolibri', 'core', 'assets', 'src', 'kolibri_module'),
       'core-base': path.join(__dirname, '..', '..', 'kolibri', 'core', 'assets', 'src', 'core-base'),
       'core-theme.styl': path.join(__dirname, '..', '..', 'kolibri', 'core', 'assets', 'src', 'core-theme.styl'),
+      'content-renderer': path.resolve('kolibri/core/assets/src/content-renderer'),
     },
     extensions: ["", ".vue", ".js"],
   },
