+++ kolibri/core/assets/src/content-renderer.vue
