+++ kolibri/plugins/audio_mp3_render/assets/src/vue/index.vue
@@ -13,6 +13,4 @@
 </script>
 
 
-<style lang="stylus" scoped>
-
-</style>
+<style lang="stylus" scoped></style>
