+++ kolibri/plugins/document_pdf_render/assets/src/vue/index.vue
@@ -19,6 +19,4 @@
 </script>
 
 
-<style lang="stylus" scoped>
-
-</style>
+<style lang="stylus" scoped></style>
