+++ kolibri/plugins/video_mp4_render/assets/src/vue/index.vue
@@ -12,6 +12,4 @@
 </script>
 
 
-<style lang="stylus" scoped>
-
-</style>
+<style lang="stylus" scoped></style>
